requirements.txt
streamlit==1.28.0
pymupdf==1.22.0
pytesseract==0.3.10
Pillow==10.0.0
